split characteristics util into multiple files
