+++ openwillis/measures/text/util/characteristics_util.py
@@ -1,33 +1,20 @@
-# author:    Vijay Yadav
+# author:    Vijay Yadav, Georgios Efstathiadis
 # website:   http://www.bklynhlth.com
 
 # import the required packages
-
 import pandas as pd
 import numpy as np
-import re
-import string
 import logging
 
 import nltk
-from vaderSentiment.vaderSentiment import SentimentIntensityAnalyzer
-from lexicalrichness import LexicalRichness
-from transformers import BertTokenizer, BertModel, BertForMaskedLM
-from sentence_transformers import SentenceTransformer
-from sklearn.metrics.pairwise import cosine_similarity
-import spacy
-import torch
+from openwillis.measures.text.util.speech.pause import get_pause_feature
+from openwillis.measures.text.util.speech.lexical import get_repetitions, get_sentiment, get_pos_tag
+from openwillis.measures.text.util.speech.coherence import get_word_coherence, get_phrase_coherence
 
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger()
 # Suppress warnings from transformers
 logging.getLogger("transformers").setLevel(logging.ERROR)
-
-# NLTK Tag list
-TAG_DICT = {"PRP": "Pronoun", "PRP$": "Pronoun", "VB": "Verb", "VBD": "Verb", "VBG": "Verb", "VBN": "Verb", "VBP": "Verb", 
-            "VBZ": "Verb", "JJ": "Adjective", "JJR": "Adjective", "JJS": "Adjective", "NN": "Noun", "NNP": "Noun", "NNS": "Noun",
-            "RB": "Adverb", "RBR": "Adverb", "RBS": "Adverb", "DT": "Determiner"}
-FIRST_PERSON_PRONOUNS = ["I", "me", "my", "mine", "myself"]
 
 def create_empty_dataframes(measures):
     """
@@ -410,1307 +397,6 @@ def filter_json_transcribe(item_data, measures):
     filter_json = pause_calculation(filter_json, measures, ['start', 'end'])
 
     return filter_json
-
-def get_num_of_syllables(text):
-    """
-    ------------------------------------------------------------------------------------------------------
-
-    This function calculates the number of syllables in the input text.
-
-    Parameters:
-    ...........
-    text: str
-        The input text.
-
-    Returns:
-    ...........
-    syllable_count: int
-        The number of syllables in the input text.
-
-    ---------------------------------------------------------------------------------------
-    """
-
-    syllable_tokenizer = nltk.tokenize.SyllableTokenizer()
-    punctuation = "!\"#$%&()*+,-./:;<=>?@[\]^_`{|}~" # remove punctuation
-    
-    syllables = [syllable_tokenizer.tokenize(token) for token in nltk.word_tokenize(text) if token not in punctuation]
-    syllable_count = sum([len(token) for token in syllables])
-
-    return syllable_count
-
-def get_pause_feature_word(word_df, df_diff, word_list, turn_index, measures):
-    """
-    ------------------------------------------------------------------------------------------------------
-
-    This function calculates various pause-related speech characteristic
-        features at the word level and adds them to the output dataframe word_df.
-
-    Parameters:
-    ...........
-    word_df: pandas dataframe
-        A dataframe containing word summary information
-    df_diff: pandas dataframe
-        A dataframe containing the word-level information
-            from the JSON response.
-    word_list: list
-        List of transcribed text at the word level.
-    turn_index: list
-        A list containing the indices of the first and last word
-    measures: dict
-        A dictionary containing the names of the columns in the output dataframes.
-
-    Returns:
-    ...........
-    word_df: pandas dataframe
-        The updated word_df dataframe.
-
-    ------------------------------------------------------------------------------------------------------
-    """
-    turn_starts = [pindex[0] for pindex in turn_index]
-    word_df[measures["word_pause"]] = df_diff[measures["pause"]].where(~df_diff[measures["old_index"]].isin(turn_starts), np.nan)
-    
-    word_df[measures["num_syllables"]] = [get_num_of_syllables(word) for word in word_list]
-    return word_df
-
-def process_pause_feature(df_diff, df, text_level, index_list, time_index, level_name, measures, language):
-    """
-    ------------------------------------------------------------------------------------------------------
-
-    This function calculates various pause-related speech
-     characteristic features at the turn
-     level and adds them to the output dataframe df.
-
-    Parameters:
-    ...........
-    df_diff: pandas dataframe
-        A dataframe containing the word-level information from the JSON response.
-    df: pandas dataframe
-        A dataframe containing turn summary information
-    text_level: list
-        List of transcribed text at the turn level.
-    index_list: list
-        A list containing the indices of the first and last word in each turn.
-    time_index: list
-        A list containing the names of the columns in json that contain
-         the start and end times of each word.
-    level_name: str
-        The name of the level being analyzed turn.
-    measures: dict
-        A dictionary containing the names of the columns in the output dataframes.
-
-    Returns:
-    ...........
-    df: pandas dataframe
-        The updated df dataframe.
-
-    ------------------------------------------------------------------------------------------------------
-    """
-
-    if level_name not in [measures["turn"]]:
-        logger.error(f"level_name must be turn")
-        return df
-
-    for j, index in enumerate(index_list):
-        try:
-            
-            rng = range(index[0], index[1] + 1)
-            level_json = df_diff[df_diff[measures["old_index"]].isin(rng)]
-            
-            pauses = level_json[measures["pause"]].values[1:]
-            level_min_val = (float(level_json.iloc[-1][time_index[1]]) - float(level_json.iloc[0][time_index[0]])) / 60
-            
-            df.loc[j, measures[f"{level_name}_minutes"]] = level_min_val
-            df.loc[j, measures[f"{level_name}_words"]] = len(level_json)
-
-            if len(pauses) == 1:
-                df.loc[j, measures["pause_var"]] = 0
-                df.loc[j, measures["pause_meandur"]] = np.mean(pauses)
-
-            elif len(pauses) > 1:
-                df.loc[j, measures["pause_var"]] = np.var(pauses)
-                df.loc[j, measures["pause_meandur"]] = np.mean(pauses)
-
-            if df.loc[j, measures[f"{level_name}_minutes"]] > 0:
-                speech_pct_val = 100 * (1 - np.sum(pauses) / (60 * df.loc[j, measures[f"{level_name}_minutes"]]))
-                df.loc[j, measures["speech_percentage"]] = speech_pct_val
-
-                if language in measures["english_langs"]:
-                    syllable_rate = (get_num_of_syllables(text_level[j]) / df.loc[j, measures[f"{level_name}_minutes"]])
-                    df.loc[j, measures["syllable_rate"]] = syllable_rate
-                
-                word_rate_val = (df.loc[j, measures[f"{level_name}_words"]] / df.loc[j, measures[f"{level_name}_minutes"]])
-                df.loc[j, measures["word_rate"]] = word_rate_val
-                
-        except Exception as e:
-            logger.error(f"Error in pause feature calculation for {level_name} {j}: {e}")
-            continue
-
-    return df
-
-def get_pause_feature_turn(turn_df, df_diff, turn_list, turn_index, time_index, measures, language):
-    """
-    ------------------------------------------------------------------------------------------------------
-
-    This function calculates various pause-related speech characteristic
-        features at the turn level and adds them to the output dataframe turn_df.
-
-    Parameters:
-    ...........
-    turn_df: pandas dataframe
-        A dataframe containing turn summary information
-    df_diff: pandas dataframe
-        A dataframe containing the word-level information
-            from the JSON response.
-    turn_list: list
-        List of transcribed text at the turn level.
-    turn_index: list
-        A list containing the indices of the first and last word
-            in each turn.
-    time_index: list
-        A list containing the names of the columns in json that contain
-            the start and end times of each word.
-    measures: dict
-        A dictionary containing the names of the columns in the output dataframes.
-
-    Returns:
-    ...........
-    turn_df: pandas dataframe
-        The updated turn_df dataframe.
-
-    ------------------------------------------------------------------------------------------------------
-    """
-    turn_starts = [uindex[0] for uindex in turn_index]
-    df_diff_turn = df_diff[df_diff[measures["old_index"]].isin(turn_starts)]
-
-    turn_df[measures["turn_pause"]] = df_diff_turn[measures["pause"]]
-    turn_df[measures["interrupt_flag"]] = False
-    
-    negative_pause = turn_df[measures["turn_pause"]] <= 0
-    turn_df.loc[negative_pause, measures["turn_pause"]] = 0
-    
-    turn_df.loc[negative_pause, measures["interrupt_flag"]] = True
-    turn_df = turn_df.reset_index(drop=True)
-
-    turn_df = process_pause_feature(df_diff, turn_df, turn_list, turn_index, time_index, measures["turn"], measures, language)
-    return turn_df
-
-def update_summ_df(df_diff, summ_df, full_text, time_index, word_df, turn_df, measures):
-    """
-    ------------------------------------------------------------------------------------------------------
-
-    This function calculates various pause-related speech characteristic
-     features at the file level and adds them to the output dataframe summ_df.
-
-    Parameters:
-    ...........
-    df_diff: pandas dataframe
-        A dataframe containing the word-level information
-         from the JSON response.
-    summ_df: pandas dataframe
-        A dataframe containing the speech characteristics of the input text.
-    time_index: list
-        A list containing the names of the columns in json
-         that contain the start and end times of each word.
-    word_df: pandas dataframe
-        A dataframe containing word summary information
-    turn_df: pandas dataframe
-        A dataframe containing turn summary information
-    measures: dict
-        A dictionary containing the names of the columns in the output dataframes.
-
-    Returns:
-    ...........
-    summ_df: pandas dataframe
-        The updated summ_df dataframe.
-
-    ------------------------------------------------------------------------------------------------------
-    """
-    if len(turn_df) > 0:
-        speech_minutes = turn_df[measures["turn_minutes"]].sum()
-        speech_words = turn_df[measures["turn_words"]].sum()
-    else:
-        speech_minutes = (float(df_diff.iloc[-1][time_index[1]]) - float(df_diff.iloc[0][time_index[0]])) / 60
-        speech_words = len(df_diff)
-
-    summ_df[measures["speech_minutes"]] = [speech_minutes]    
-    summ_df[measures["speech_words"]] = [speech_words]
-
-    if speech_minutes > 0:
-        
-        summ_df[measures["word_rate"]] = (summ_df[measures["speech_words"]] / summ_df[measures["speech_minutes"]])
-        summ_df[measures["syllable_rate"]] = (get_num_of_syllables(full_text) / summ_df[measures["speech_minutes"]])
-        summ_df[measures["speech_percentage"]] = 100 * (summ_df[measures["speech_minutes"]] / summ_df[measures["file_length"]])
-
-    if len(word_df[measures["word_pause"]]) > 1:
-        summ_df[measures["word_pause_mean"]] = word_df[measures["word_pause"]].mean(skipna=True)
-        summ_df[measures["word_pause_var"]] = word_df[measures["word_pause"]].var(skipna=True)
-    
-    if len(turn_df) > 0:
-        summ_df[measures["num_turns"]] = len(turn_df)
-        summ_df[measures["turn_minutes_mean"]] = turn_df[measures["turn_minutes"]].mean(skipna=True)
-        
-        summ_df[measures["turn_words_mean"]] = turn_df[measures["turn_words"]].mean(skipna=True)
-        summ_df[measures["turn_pause_mean"]] = turn_df[measures["turn_pause"]].mean(skipna=True)
-        
-        summ_df[measures["num_one_word_turns"]] = len(turn_df[turn_df[measures["turn_words"]] == 1])
-        summ_df[measures["num_interrupts"]] = len(turn_df[turn_df[measures["interrupt_flag"]]==True])
-
-    return summ_df
-
-def get_pause_feature(json_conf, df_list, text_list, turn_index, measures, time_index, language):
-    """
-    ------------------------------------------------------------------------------------------------------
-
-    This function calculates various pause-related
-     speech characteristic features
-
-    Parameters:
-    ...........
-    json_conf: list
-        JSON response object.
-    df_list: list
-        List of pandas dataframes: word_df, turn_df, summ_df
-    text_list: list
-        List of transcribed text: split into words, turns, and full text.
-    turn_index: list
-        List of indices for text_list.
-    measures: dict
-        A dictionary containing the names of the columns in the output dataframes.
-    time_index: list
-        timepoint index (start/end)
-    language: str
-        Language of the transcribed text.
-
-    Returns:
-    ...........
-    df_feature: list
-        List of updated pandas dataframes (word_df, turn_df and summ_df)
-
-    ------------------------------------------------------------------------------------------------------
-    """
-    if len(json_conf) <= 0:
-        return df_list
-
-    word_df, turn_df, summ_df = df_list
-    word_list, turn_list, full_text = text_list
-    df_diff = pd.DataFrame(json_conf)
-
-    # Calculate the pause time between; each word and add the results to pause_list
-    if measures["pause"] not in df_diff.columns:
-        df_diff[measures["pause"]] = df_diff[time_index[0]].astype(float) - df_diff[time_index[1]].astype(float).shift(1)
-
-    # word-level analysis
-    word_df = get_pause_feature_word(word_df, df_diff, word_list, turn_index, measures)
-
-    # turn-level analysis
-    if len(turn_index) > 0:
-        turn_df = get_pause_feature_turn(turn_df, df_diff, turn_list, turn_index, time_index, measures, language)
-
-    # file-level analysis
-    summ_df = update_summ_df(df_diff, summ_df, full_text, time_index, word_df, turn_df, measures)
-    df_feature = [word_df, turn_df, summ_df]
-    return df_feature
-
-def get_mattr(text, lemmatizer, window_size=50):
-    """
-    ------------------------------------------------------------------------------------------------------
-    This function calculates the Moving Average Type-Token Ratio (MATTR)
-     of the input text using the
-     LexicalRichness library.
-
-    Parameters:
-    ...........
-    text : str
-        The input text to be analyzed.
-    lemmatizer : spacy lemmatizer
-        The lemmatizer to be used in the calculation.
-    window_size : int
-        The size of the window to be used in the calculation.
-
-    Returns:
-    ...........
-    mattr : float
-        The calculated MATTR value.
-
-    ------------------------------------------------------------------------------------------------------
-    """
-
-    words = nltk.word_tokenize(text)
-    words = [w.translate(str.maketrans('', '', string.punctuation)).lower() for w in words]
-    words = [w for w in words if w != '']
-    words_texts = [token.lemma_ for token in lemmatizer(' '.join(words))]
-    filter_punc = " ".join(words_texts)
-    mattr = np.nan
-
-    lex_richness = LexicalRichness(filter_punc)
-    if lex_richness.words > 0:
-        mattr = lex_richness.mattr(window_size=min(window_size, lex_richness.words))
-
-    return mattr
-
-def get_tag(word_df, word_list, measures):
-    """
-    ------------------------------------------------------------------------------------------------------
-
-    This function performs part-of-speech
-     tagging on the input text using NLTK, and returns
-     word-level part-of-speech tags.
-
-    Parameters:
-    ...........
-    word_df: pandas dataframe
-        A dataframe containing word summary information.
-    word_list: list
-        List of transcribed text at the word level.
-    measures: dict
-        A dictionary containing the names of the columns in the output dataframes.
-
-    Returns:
-    ...........
-    word_df: pandas dataframe
-        The updated word_df dataframe.
-
-    ------------------------------------------------------------------------------------------------------
-    """
-    tag_list = nltk.pos_tag(word_list)
-    
-    tag_list_pos = [TAG_DICT[tag[1]] if tag[1] in TAG_DICT.keys() else "Other" for tag in tag_list]
-    word_df[measures["part_of_speech"]] = tag_list_pos
-
-    word_df[measures["first_person"]] = [word in FIRST_PERSON_PRONOUNS for word in word_list]
-    # make non pronouns NaN
-    word_df[measures["first_person"]] = word_df[measures["first_person"]].where(word_df[measures["part_of_speech"]] == "Pronoun", np.nan)
-
-    present_tense = ["VBP", "VBZ"]
-    past_tense = ["VBD", "VBN"]
-    tag_list_verb = ["Present" if tag[1] in present_tense else "Past" if tag[1] in past_tense else "Other" for tag in tag_list]
-    word_df[measures["verb_tense"]] = tag_list_verb
-    # make non verbs NaN
-    word_df[measures["verb_tense"]] = word_df[measures["verb_tense"]].where(word_df[measures["part_of_speech"]] == "Verb", np.nan)
-
-    return word_df
-
-def calculate_first_person_sentiment(df, measures):
-    """
-    ------------------------------------------------------------------------------------------------------
-
-    This function calculates a measure of the influence of sentiment on the use of first person pronouns.
-
-    Parameters:
-    ...........
-    df: pandas dataframe
-        A dataframe containing summary information.
-    measures: dict
-        A dictionary containing the names of the columns in the output dataframes.
-
-    Returns:
-    ...........
-    res1: list
-        A list containing the calculated measure of the influence of positive sentiment on the use of first person pronouns.
-    res2: list
-        A list containing the calculated measure of the influence of negative sentiment on the use of first person pronouns.
-
-    ------------------------------------------------------------------------------------------------------
-    """
-    
-    res1 = []
-    res2 = []
-    for i in range(len(df)):
-        perc = df.loc[i, measures["first_person_percentage"]]
-        pos = df.loc[i, measures["pos"]]
-        neg = df.loc[i, measures["neg"]]
-
-        if perc is np.nan:
-            res1.append(np.nan)
-            res2.append(np.nan)
-            continue
-
-        res1.append((100-perc)*pos)
-        res2.append(perc*neg)
-
-    return res1, res2
-
-def calculate_first_person_percentage(text):
-    """
-    ------------------------------------------------------------------------------------------------------
-
-    This function calculates the percentage of first person pronouns in the input text.
-
-    Parameters:
-    ...........
-    text: str
-        The input text to be analyzed.
-
-    Returns:
-    ...........
-    float
-        The calculated percentage of first person pronouns in the input text.
-
-    ------------------------------------------------------------------------------------------------------
-    """
-
-    words = nltk.word_tokenize(text)
-    tags = nltk.pos_tag(words)
-    # filter out non pronouns
-    pronouns = [tag[0] for tag in tags if tag[1] == "PRP" or tag[1] == "PRP$"]
-    if len(pronouns) == 0:
-        return np.nan
-
-    first_person_pronouns = len([word for word in pronouns if word in FIRST_PERSON_PRONOUNS])
-    return (first_person_pronouns / len(pronouns)) * 100
-
-def get_first_person_turn(turn_df, turn_list, measures):
-    """
-    ------------------------------------------------------------------------------------------------------
-
-    This function calculates measures related to the first person pronouns in each turn.
-     Specifically, it calculates the percentage of first person pronouns in each turn,
-     and the influence of sentiment on the use of first person pronouns.
-
-    Parameters:
-    ...........
-    turn_df: pandas dataframe
-        A dataframe containing turn summary information.
-    turn_list: list
-        List of transcribed text at the turn level.
-
-    Returns:
-    ...........
-    turn_df: pandas dataframe
-        The updated turn_df dataframe.
-
-    ------------------------------------------------------------------------------------------------------
-    """
-    first_person_percentages = [calculate_first_person_percentage(turn) for turn in turn_list]
-
-    turn_df[measures["first_person_percentage"]] = first_person_percentages
-
-    first_pos, first_neg = calculate_first_person_sentiment(turn_df, measures)
-
-    turn_df[measures["first_person_sentiment_positive"]] = first_pos
-    turn_df[measures["first_person_sentiment_negative"]] = first_neg
-
-    return turn_df
-
-def get_first_person_summ(summ_df, turn_df, full_text, measures):
-    """
-    ------------------------------------------------------------------------------------------------------
-
-    This function calculates measures related to the first person pronouns in the transcript.
-
-    Parameters:
-    ...........
-    summ_df: pandas dataframe
-        A dataframe containing summary information.
-    turn_df: pandas dataframe
-        A dataframe containing turn summary information.
-    full_text: str
-        The full transcribed text.
-    measures: dict
-        A dictionary containing the names of the columns in the output dataframes.
-
-    Returns:
-    ...........
-    summ_df: pandas dataframe
-        The updated summ_df dataframe.
-
-    ------------------------------------------------------------------------------------------------------
-    """
-
-    summ_df[measures["first_person_percentage"]] = calculate_first_person_percentage(full_text)
-
-    if len(turn_df) > 0:
-        summ_df[measures["first_person_sentiment_positive"]] = turn_df[measures["first_person_sentiment_positive"]].mean(skipna=True)
-        summ_df[measures["first_person_sentiment_negative"]] = turn_df[measures["first_person_sentiment_negative"]].mean(skipna=True)
-
-        first_person_sentiment = []
-        for i in range(len(turn_df)):
-            if turn_df.loc[i, measures["pos"]] > turn_df.loc[i, measures["neg"]]:
-                first_person_sentiment.append(turn_df.loc[i, measures["first_person_sentiment_positive"]])
-            else:    
-                first_person_sentiment.append(turn_df.loc[i, measures["first_person_sentiment_negative"]])
-
-        summ_df[measures["first_person_sentiment_overall"]] = np.nanmean(first_person_sentiment)
-    else:
-        first_pos, first_neg = calculate_first_person_sentiment(summ_df, measures)
-        summ_df[measures["first_person_sentiment_positive"]] = first_pos
-        summ_df[measures["first_person_sentiment_negative"]] = first_neg
-
-        if summ_df[measures["pos"]].values[0] > summ_df[measures["neg"]].values[0]:
-            summ_df[measures["first_person_sentiment_overall"]] = summ_df[measures["first_person_sentiment_positive"]].values[0]
-        else:
-            summ_df[measures["first_person_sentiment_overall"]] = summ_df[measures["first_person_sentiment_negative"]].values[0]
-
-    return summ_df
-
-def get_pos_tag(df_list, text_list, measures):
-    """
-    ------------------------------------------------------------------------------------------------------
-
-    This function calculates the part-of-speech measures
-        and adds them to the output dataframes
-
-    Parameters:
-    ...........
-    df_list: list
-        List of pandas dataframes.
-    text_list: list
-        List of transcribed text.
-    measures: dict
-        A dictionary containing the names of the columns in the output dataframes.
-
-    Returns:
-    ...........
-    df_list: list
-        List of updated pandas dataframes.
-
-    ------------------------------------------------------------------------------------------------------
-    """
-    word_df, turn_df, summ_df = df_list
-    word_list, turn_list, full_text = text_list
-
-    word_df = get_tag(word_df, word_list, measures)
-
-    if len(turn_list) > 0:
-        turn_df = get_first_person_turn(turn_df, turn_list, measures)
-
-    summ_df = get_first_person_summ(summ_df, turn_df, full_text, measures)
-
-    df_list = [word_df, turn_df, summ_df]
-    return df_list
-
-def get_sentiment(df_list, text_list, measures):
-    """
-    ------------------------------------------------------------------------------------------------------
-
-    This function calculates the sentiment scores of the input text using
-     VADER, and adds them to the output dataframe summ_df.
-
-    Parameters:
-    ...........
-    df_list: list
-        List of pandas dataframes.
-    text_list: list
-        List of transcribed text.
-    measures: dict
-        A dictionary containing the names of the columns in the output dataframes.
-
-    Returns:
-    ...........
-    df_list: list
-        List of updated pandas dataframes.
-
-    ------------------------------------------------------------------------------------------------------
-    """
-    word_df, turn_df, summ_df = df_list
-    _, turn_list, full_text = text_list
-    lemmatizer = spacy.load('en_core_web_sm')
-
-    sentiment = SentimentIntensityAnalyzer()
-    cols = [measures["neg"], measures["neu"], measures["pos"], measures["compound"], measures["speech_mattr_5"], measures["speech_mattr_10"], measures["speech_mattr_25"], measures["speech_mattr_50"], measures["speech_mattr_100"]]
-
-    for idx, u in enumerate(turn_list):
-        try:
-            
-            sentiment_dict = sentiment.polarity_scores(u)
-            mattrs = [get_mattr(u, lemmatizer, window_size=ws) for ws in [5, 10, 25, 50, 100]]
-            turn_df.loc[idx, cols] = list(sentiment_dict.values()) + mattrs
-            
-        except Exception as e:
-            logger.error(f"Error in sentiment analysis: {e}")
-            continue
-            
-    sentiment_dict = sentiment.polarity_scores(full_text)
-    mattrs = [get_mattr(full_text, lemmatizer, window_size=ws) for ws in [5, 10, 25, 50, 100]]
-
-    summ_df.loc[0, cols] = list(sentiment_dict.values()) + mattrs
-    df_list = [word_df, turn_df, summ_df]
-    return df_list
-
-def calculate_repetitions(words_texts, phrases_texts):
-    """
-    ------------------------------------------------------------------------------------------------------
-
-    This function calculates the percentage of repeated words and phrases in the input lists.
-
-    Parameters:
-    ...........
-    words_texts: list
-        List of transcribed text at the word level.
-    phrases_texts: list
-        List of transcribed text at the phrase level.
-
-    Returns:
-    ...........
-    word_reps_perc: float
-        The percentage of repeated words in the input lists.
-    phrase_reps_perc: float
-        The percentage of repeated phrases in the input lists.
-
-    ------------------------------------------------------------------------------------------------------
-    """
-    # remove punctuation - lowercase
-    words_texts = [word.translate(str.maketrans('', '', string.punctuation)).lower() for word in words_texts]
-    phrases_texts = [phrase.translate(str.maketrans('', '', string.punctuation)).lower() for phrase in phrases_texts]
-
-    # remove empty strings
-    words_texts = [word for word in words_texts if word != '']
-    phrases_texts = [phrase for phrase in phrases_texts if phrase != '']
-
-    # sliding window of 10 words for repetitions
-    if len(words_texts) <= 10:
-        word_reps = len(words_texts) - len(set(words_texts))
-        word_perc = 100*word_reps/len(words_texts)
-    else:
-        word_reps_list = []
-        for i in range(len(words_texts) - 9):
-            window = words_texts[i:i+10]
-            word_reps_list.append(100*(len(window) - len(set(window))) / len(window))
-        word_perc = np.mean(word_reps_list)
-
-    if len(phrases_texts) == 0:
-        return word_perc, np.nan
-    
-    # sliding window of 3 phrases for repetitions
-    if len(phrases_texts) <= 3:
-        phrase_reps = len(phrases_texts) - len(set(phrases_texts))
-        phrase_perc = 100*phrase_reps/len(phrases_texts)
-    else:
-        phrase_reps_list = []
-        for i in range(len(phrases_texts) - 3):
-            window = phrases_texts[i:i+3]
-            phrase_reps_list.append(100*(len(window) - len(set(window))) / len(window))
-        phrase_perc = np.mean(phrase_reps_list)
-
-    return word_perc, phrase_perc
-
-
-def get_repetitions(df_list, utterances_speaker, utterances_speaker_filtered, language, measures):
-    """
-
-    This function calculates the percentage of repeated words and phrases in the input text
-    and adds them to the output dataframes.
-
-    Parameters:
-    ...........
-    df_list: list
-        List of pandas dataframes.
-    utterances_speaker: pandas dataframe
-        A dataframe containing the turns extracted from the JSON object for the specified speaker.
-    utterances_speaker_filtered: pandas dataframe
-        A dataframe containing the turns extracted from the JSON object for the specified speaker
-        after filtering out turns with less than min_turn_length words.
-    language: str
-        Language of the transcribed text.
-    measures: dict
-        A dictionary containing the names of the columns in the output dataframes.
-
-    Returns:
-    ...........
-    df_list: list
-        List of updated pandas dataframes.
-
-    """
-    
-    word_df, turn_df, summ_df = df_list
-
-    # turn-level
-    if len(turn_df) > 0:
-        for i in range(len(utterances_speaker_filtered)):
-            row = utterances_speaker_filtered.iloc[i]
-            words_texts = row[measures['words_texts']]
-            phrases_texts = row[measures['phrases_texts']]
-
-            word_reps_perc, phrase_reps_perc = calculate_repetitions(words_texts, phrases_texts)
-
-            turn_df.loc[i, measures['word_repeat_percentage']] = word_reps_perc
-            turn_df.loc[i, measures['phrase_repeat_percentage']] = phrase_reps_perc
-
-    # summary-level
-    if len(turn_df) > 0:
-        summ_df[measures['word_repeat_percentage']] = turn_df[measures['word_repeat_percentage']].mean(skipna=True)
-        summ_df[measures['phrase_repeat_percentage']] = turn_df[measures['phrase_repeat_percentage']].mean(skipna=True)
-    else:
-        words_texts = [word for words in utterances_speaker[measures['words_texts']] for word in words]
-        phrases_texts = [phrase for phrases in utterances_speaker[measures['phrases_texts']] for phrase in phrases]
-
-        word_reps_perc, phrase_reps_perc = calculate_repetitions(words_texts, phrases_texts)
-
-        summ_df[measures['word_repeat_percentage']] = word_reps_perc
-        summ_df[measures['phrase_repeat_percentage']] = phrase_reps_perc
-
-    df_list = [word_df, turn_df, summ_df]
-    return df_list
-
-def get_word_embeddings(word_list, tokenizer, model):
-    """
-    ------------------------------------------------------------------------------------------------------
-
-    This function calculates the word embeddings for the input text using BERT.
-
-    Parameters:
-    ...........
-    word_list: list
-        List of transcribed text at the word level.
-    language: str
-        Language of the transcribed text.
-
-    Returns:
-    ...........
-    word_embeddings: numpy array
-        The calculated word embeddings.
-
-    ------------------------------------------------------------------------------------------------------
-    """
-    if len(word_list) >= 512:
-        # split the text into chunks of 512 tokens
-        word_list = [word_list[i:i+512] for i in range(0, len(word_list), 512)]
-        word_embeddings = []
-        for chunk in word_list:
-            inputs = tokenizer(chunk, return_tensors='pt', padding=True)
-            outputs = model(**inputs)
-            word_embeddings.append(outputs.last_hidden_state.mean(1).detach().numpy())
-        word_embeddings = np.concatenate(word_embeddings, axis=0)
-    else:
-        inputs = tokenizer(word_list, return_tensors='pt', padding=True)
-        outputs = model(**inputs)
-        # Average pooling of the hidden states
-        word_embeddings = outputs.last_hidden_state.mean(1).detach().numpy()
-    return word_embeddings
-
-def get_word_coherence_utterance(row, tokenizer, model, measures):
-    """
-    ------------------------------------------------------------------------------------------------------
-
-    This function calculates word coherence measures for a single utterance.
-
-    Parameters:
-    ...........
-    row: pandas dataframe
-        A dataframe containing the turns extracted from the JSON object for the specified speaker.
-    tokenizer: BertTokenizer
-        A tokenizer object for BERT.
-    model: BertModel
-        A BERT model object.
-    measures: dict
-        A dictionary containing the names of the columns in the output dataframes.
-
-    Returns:
-    ...........
-    word_coherence: list
-        A list containing the calculated semantic similarity of each word to the immediately preceding word.
-    word_coherence_5: list
-        A list containing the calculated semantic similarity of each word in 5-words window.
-    word_coherence_10: list
-        A list containing the calculated semantic similarity of each word in 10-words window.
-    word_word_variability: dict
-        A dictionary containing the calculated word-to-word variability at k inter-word distances.
-
-    ------------------------------------------------------------------------------------------------------
-    """
-    words_texts = row[measures['words_texts']]
-    if len(words_texts) == 0:
-        # return empty lists if no words in the utterance
-        return [np.nan]*len(words_texts), [np.nan]*len(words_texts), [np.nan]*len(words_texts), {k: [np.nan]*len(words_texts) for k in range(2, 11)}
-
-    word_embeddings = get_word_embeddings(words_texts, tokenizer, model)
-    similarity_matrix = cosine_similarity(word_embeddings)
-
-    # calculate semantic similarity of each word to the immediately preceding word
-    if len(words_texts) > 1:
-        word_coherence = [np.nan] + [similarity_matrix[j, j-1] for j in range(1, len(words_texts))]
-    else:
-        word_coherence = [np.nan]*len(words_texts)
-
-    # calculate semantic similarity of each word in 5-words window
-    if len(words_texts) > 5:
-        word_coherence_5 = [np.nan]*2 + [np.mean(similarity_matrix[j-2:j+3, j]) for j in range(2, len(words_texts)-2)] + [np.nan]*2
-    else:
-        word_coherence_5 = [np.nan]*len(words_texts)
-
-    # calculate semantic similarity of each word in 10-words window
-    if len(words_texts) > 10:
-        word_coherence_10 = [np.nan]*5 + [np.mean(similarity_matrix[j-5:j+6, j]) for j in range(5, len(words_texts)-5)] + [np.nan]*5
-    else:
-        word_coherence_10 = [np.nan]*len(words_texts)
-
-    # calculate word-to-word variability at k inter-word distances (for k from 2 to 10)
-    # indicating semantic similarity between each word and the next following word at k inter-word distance
-    word_word_variability = {}
-    for k in range(2, 11):
-        if len(words_texts) > k:
-            word_word_variability[k] = [similarity_matrix[j, j+k] for j in range(len(words_texts)-k)] + [np.nan]*k
-        else:
-            word_word_variability[k] = [np.nan]*len(words_texts)
-
-    return word_coherence, word_coherence_5, word_coherence_10, word_word_variability    
-
-def get_word_coherence_summary(word_df, summ_df, measures):
-    """
-    ------------------------------------------------------------------------------------------------------
-
-    This function summarizes the word coherence measures at the summary level.
-
-    Parameters:
-    ...........
-    word_df: pandas dataframe
-        A dataframe containing word summary information.
-    summ_df: pandas dataframe
-        A dataframe containing summary information.
-    measures: dict
-        A dictionary containing the names of the columns in the output dataframes.
-
-    Returns:
-    ...........
-    summ_df: pandas dataframe
-        The updated summ_df dataframe.
-
-    ------------------------------------------------------------------------------------------------------
-    """
-
-    summ_df[measures['word_coherence_mean']] = word_df[measures['word_coherence']].mean(skipna=True)
-    summ_df[measures['word_coherence_var']] = word_df[measures['word_coherence']].var(skipna=True)
-    summ_df[measures['word_coherence_5_mean']] = word_df[measures['word_coherence_5']].mean(skipna=True)
-    summ_df[measures['word_coherence_5_var']] = word_df[measures['word_coherence_5']].var(skipna=True)
-    summ_df[measures['word_coherence_10_mean']] = word_df[measures['word_coherence_10']].mean(skipna=True)
-    summ_df[measures['word_coherence_10_var']] = word_df[measures['word_coherence_10']].var(skipna=True)
-    for k in range(2, 11):
-        summ_df[measures[f'word_coherence_variability_{k}_mean']] = word_df[measures[f'word_coherence_variability_{k}']].mean(skipna=True)
-        summ_df[measures[f'word_coherence_variability_{k}_var']] = word_df[measures[f'word_coherence_variability_{k}']].var(skipna=True)
-
-    return summ_df
-
-def get_word_coherence(df_list, utterances_speaker, min_coherence_turn_length, language, measures):
-    """
-    ------------------------------------------------------------------------------------------------------
-
-    This function calculates word coherence measures
-
-    Parameters:
-    ...........
-    df_list: list
-        List of pandas dataframes.
-    utterances_speaker: pandas dataframe
-        A dataframe containing the turns extracted from the JSON object for the specified speaker.
-    min_coherence_turn_length: int
-        Minimum number of words in a turn for word coherence analysis.
-    language: str
-        Language of the transcribed text.
-    measures: dict
-        A dictionary containing the names of the columns in the output dataframes.
-
-    Returns:
-    ...........
-    df_list: list
-        List of updated pandas dataframes.
-
-    ------------------------------------------------------------------------------------------------------
-    """
-    word_df, turn_df, summ_df = df_list
-
-    # model init
-    if language in measures["english_langs"]:
-        tokenizer = BertTokenizer.from_pretrained('bert-base-cased')
-        model = BertModel.from_pretrained('bert-base-cased')
-    elif language in measures["supported_langs_bert"]:
-        tokenizer = BertTokenizer.from_pretrained('bert-base-multilingual-cased')
-        model = BertModel.from_pretrained('bert-base-multilingual-cased')
-    else:
-        logger.error(f"Language {language} not supported for word coherence analysis")
-        return df_list
-
-
-    # word-level
-    word_coherence_list = []
-    word_coherence_5_list = []
-    word_coherence_10_list = []
-    word_coherence_variability = {
-        x: [] for x in range(2, 11)
-    }
-    for i in range(len(utterances_speaker)):
-        row = utterances_speaker.iloc[i]
-
-        if len(row[measures['words_texts']]) < min_coherence_turn_length:
-            word_coherence_list += [np.nan]*len(row[measures['words_texts']])
-            word_coherence_5_list += [np.nan]*len(row[measures['words_texts']])
-            word_coherence_10_list += [np.nan]*len(row[measures['words_texts']])
-            for k in range(2, 11):
-                word_coherence_variability[k] += [np.nan]*len(row[measures['words_texts']])
-            continue
-
-        word_coherence, word_coherence_5, word_coherence_10, word_word_variability = get_word_coherence_utterance(row, tokenizer, model, measures)
-
-        word_coherence_list += word_coherence
-        word_coherence_5_list += word_coherence_5
-        word_coherence_10_list += word_coherence_10
-        for k in range(2, 11):
-            word_coherence_variability[k] += word_word_variability[k]
-    
-    word_df[measures['word_coherence']] = word_coherence_list
-    word_df[measures['word_coherence_5']] = word_coherence_5_list
-    word_df[measures['word_coherence_10']] = word_coherence_10_list
-    for k in range(2, 11):
-        word_df[measures[f'word_coherence_variability_{k}']] = word_coherence_variability[k]
-
-    # summary-level
-    summ_df = get_word_coherence_summary(word_df, summ_df, measures)
-
-    df_list = [word_df, turn_df, summ_df]
-    return df_list
-
-def calculate_perplexity(text, model, tokenizer):
-    """
-    ------------------------------------------------------------------------------------------------------
-
-    This function calculates the pseudo-perplexity of the input text using BERT.
-
-    Parameters:
-    ...........
-    text: str
-        The input text to be analyzed.
-    model: BertForMaskedLM
-        A BERT model.
-    tokenizer: BertTokenizer
-        A BERT tokenizer.
-
-    Returns:
-    ...........
-    Tuple of floats
-        The calculated pseudo-perplexity of the input text.
-        The calculated pseudo-perplexity of the input text using 2 words before and after the masked token.
-        The calculated pseudo-perplexity of the input text using 5 words before and after the masked token.
-        The calculated pseudo-perplexity of the input text using 7 words before and after the masked token.
-
-    ------------------------------------------------------------------------------------------------------
-    """
-    # Tokenize input text
-    clean_text = text.translate(str.maketrans('', '', string.punctuation))
-    clean_text = re.sub(r'\s+', ' ', clean_text).strip()
-    if len(clean_text) == 0 or len(clean_text.split()) < 2:
-        return np.nan
-
-    tokens = tokenizer(clean_text, return_tensors='pt')
-    input_ids = tokens.input_ids
-    masked_input_ids = input_ids.clone()
-
-    log_probs = []
-    log_probs_5 = []
-    log_probs_11 = []
-    log_probs_15 = []
-    # Iterate over each token in the input
-    for i in range(input_ids.size(1)):
-        # filter so that all input_ids include the masked location +- 256 tokens
-        if i < 256:
-            start = 0
-        else:
-            start = i - 256
-        if i > input_ids.size(1) - 256:
-            end = input_ids.size(1)
-        else:
-            end = i + 256
-
-        masked_input_ids[0, i] = tokenizer.mask_token_id
-
-        input_ids_filtered = input_ids[:, start:(end+1)]
-        masked_input_ids_filtered = masked_input_ids[:, start:(end+1)]
-        # get new i from that
-        idx = i - start
-
-        # use only 2 words before and after the masked token
-        if i < 2:
-            start_5 = 0
-        else:
-            start_5 = i - 2
-        if i > input_ids.size(1) - 2:
-            end_5 = input_ids.size(1)
-        else:
-            end_5 = i + 2
-
-        input_ids_filtered_5 = input_ids[:, start_5:(end_5+1)]
-        masked_input_ids_filtered_5 = masked_input_ids[:, start_5:(end_5+1)]
-        idx_5 = i - start_5
-
-        # use only 5 words before and after the masked token
-        if i < 5:
-            start_11 = 0
-        else:
-            start_11 = i - 5
-        if i > input_ids.size(1) - 5:
-            end_11 = input_ids.size(1)
-        else:
-            end_11 = i + 5
-
-        input_ids_filtered_11 = input_ids[:, start_11:(end_11+1)]
-        masked_input_ids_filtered_11 = masked_input_ids[:, start_11:(end_11+1)]
-        idx_11 = i - start_11
-
-        # use only 7 words before and after the masked token
-        if i < 7:
-            start_15 = 0
-        else:
-            start_15 = i - 7
-        if i > input_ids.size(1) - 7:
-            end_15 = input_ids.size(1)
-        else:
-            end_15 = i + 7
-
-        input_ids_filtered_15 = input_ids[:, start_15:(end_15+1)]
-        masked_input_ids_filtered_15 = masked_input_ids[:, start_15:(end_15+1)]
-        idx_15 = i - start_15
-
-        with torch.no_grad():
-            outputs = model(input_ids=masked_input_ids_filtered, labels=input_ids_filtered)
-            outputs_5 = model(input_ids=masked_input_ids_filtered_5, labels=input_ids_filtered_5)
-            outputs_11 = model(input_ids=masked_input_ids_filtered_11, labels=input_ids_filtered_11)
-            outputs_15 = model(input_ids=masked_input_ids_filtered_15, labels=input_ids_filtered_15)
-
-        # Calculate log probability of the original token
-        logit_prob = outputs.logits[0, idx].softmax(dim=0)
-        true_log_prob = logit_prob[input_ids[0, i]].log().item()
-        log_probs.append(true_log_prob)
-
-        logit_prob_5 = outputs_5.logits[0, idx_5].softmax(dim=0)
-        true_log_prob_5 = logit_prob_5[input_ids[0, i]].log().item()
-        log_probs_5.append(true_log_prob_5)
-
-        logit_prob_11 = outputs_11.logits[0, idx_11].softmax(dim=0)
-        true_log_prob_11 = logit_prob_11[input_ids[0, i]].log().item()
-        log_probs_11.append(true_log_prob_11)
-
-        logit_prob_15 = outputs_15.logits[0, idx_15].softmax(dim=0)
-        true_log_prob_15 = logit_prob_15[input_ids[0, i]].log().item()
-        log_probs_15.append(true_log_prob_15)
-
-        # Unmask the token for the next iteration
-        masked_input_ids[0, i] = input_ids[0, i]
-
-    # Calculate perplexity
-    perplexity = np.exp(-np.mean(log_probs))
-    perplexity_5 = np.exp(-np.mean(log_probs_5))
-    perplexity_11 = np.exp(-np.mean(log_probs_11))
-    perplexity_15 = np.exp(-np.mean(log_probs_15))
-    return perplexity, perplexity_5, perplexity_11, perplexity_15
-
-def calculate_phrase_tangeniality(phrases_texts, utterance_text, sentence_encoder, bert, tokenizer):
-    """
-    ------------------------------------------------------------------------------------------------------
-
-    This function calculates the semantic similarity of each phrase to the immediately preceding phrase,
-    the semantic similarity of each phrase to the phrase 2 turns before, and the pseudo-perplexity of the turn.
-
-    Parameters:
-    ...........
-    phrases_texts: list
-        List of transcribed text at the phrase level.
-    utterance_text: str
-        The full transcribed text.
-    sentence_encoder: SentenceTransformer
-        A SentenceTransformer model.
-    bert: BertForMaskedLM
-        A BERT model.
-    tokenizer: BertTokenizer
-        A BERT tokenizer.
-
-    Returns:
-    ...........
-    sentence_tangeniality1: float
-        The semantic similarity of each phrase to the immediately preceding phrase.
-    sentence_tangeniality2: float
-        The semantic similarity of each phrase to the phrase 2 turns before.
-    perplexity: float
-        The pseudo-perplexity of the turn.
-    perplexity_5: float
-        The pseudo-perplexity of the turn using 2 words before and after the masked token.
-    perplexity_11: float
-        The pseudo-perplexity of the turn using 5 words before and after the masked token.
-    perplexity_15: float
-        The pseudo-perplexity of the turn using 7 words before and after the masked token.
-
-    ------------------------------------------------------------------------------------------------------
-    """
-    if sentence_encoder is not None and len(phrases_texts) > 0:
-        phrase_embeddings = sentence_encoder.encode(phrases_texts)
-        similarity_matrix = cosine_similarity(phrase_embeddings)
-
-        # calculate semantic similarity of each phrase to the immediately preceding phrase
-        if len(phrases_texts) > 1:
-            sentence_tangeniality1 = np.mean([similarity_matrix[j-1, j] for j in range(1, len(phrases_texts))])
-        else:
-            sentence_tangeniality1 = np.nan
-
-        # calculate semantic similarity of each phrase to the phrase 2 turns before
-        if len(phrases_texts) > 2:
-            sentence_tangeniality2 = np.mean([similarity_matrix[j-2, j] for j in range(2, len(phrases_texts))])
-        else:
-            sentence_tangeniality2 = np.nan
-    else:
-        sentence_tangeniality1 = np.nan
-        sentence_tangeniality2 = np.nan
-
-    if tokenizer is not None and bert is not None:
-        # calculate pseudo-perplexity of the turn and indicating how predictable the turn is
-        perplexity, perplexity_5, perplexity_11, perplexity_15 = calculate_perplexity(utterance_text, bert, tokenizer)
-    else:
-        perplexity, perplexity_5, perplexity_11, perplexity_15 = np.nan, np.nan, np.nan
-
-    return sentence_tangeniality1, sentence_tangeniality2, perplexity, perplexity_5, perplexity_11, perplexity_15
-
-def calculate_slope(y):
-    """
-    ------------------------------------------------------------------------------------------------------
-    This function calculates the slope
-     of the input list using linear regression
-
-    Parameters:
-    ...........
-    y: list
-        A list of values.
-
-    Returns:
-    ...........
-    float
-        The calculated slope of the input list.
-
-    ------------------------------------------------------------------------------------------------------
-    """
-    # remove NaNs
-    y = [val for val in y if not np.isnan(val)]
-
-    x = range(len(y))
-    try:
-        slope, _ = np.polyfit(x, y, 1)
-    except:
-        slope = np.nan
-
-    return slope
-
-
-def get_phrase_coherence(df_list, utterances_filtered, min_coherence_turn_length, speaker_label, language, measures):
-    """
-    ------------------------------------------------------------------------------------------------------
-
-    This function calculates turn coherence measures
-
-    Parameters:
-    ...........
-    df_list: list
-        List of pandas dataframes.
-    utterances_filtered: pandas dataframe
-        A dataframe containing the turns extracted from the JSON object
-        after filtering out turns with less than min_turn_length words of the specified speaker.
-    min_coherence_turn_length: int
-        Minimum number of words in a turn for word coherence analysis.
-    speaker_label: str
-        Speaker label
-    language: str
-        Language of the transcribed text.
-    measures: dict
-        A dictionary containing the names of the columns in the output dataframes.
-
-    Returns:
-    ...........
-    df_list: list
-        List of updated pandas dataframes.
-
-    ------------------------------------------------------------------------------------------------------
-    """
-    word_df, turn_df, summ_df = df_list
-
-    if language in measures["english_langs"]:
-        sentence_encoder = SentenceTransformer('sentence-transformers/all-MiniLM-L6-v2')
-
-        tokenizer = BertTokenizer.from_pretrained('bert-base-cased')
-        bert = BertForMaskedLM.from_pretrained('bert-base-cased')
-    else:
-        if language not in measures["supported_langs_sentence_embeddings"] + measures["supported_langs_bert"]:
-            logger.error(f"Language {language} not supported for phrase coherence nor perplexity analysis")
-            return df_list
-
-        if language in measures["supported_langs_sentence_embeddings"]:
-            sentence_encoder = SentenceTransformer('sentence-transformers/paraphrase-multilingual-MiniLM-L12-v2')
-        else:
-            sentence_encoder = None
-            logger.error(f"Language {language} not supported for phrase coherence analysis")
-
-        if language in measures["supported_langs_bert"]:
-            tokenizer = BertTokenizer.from_pretrained('bert-base-multilingual-cased')
-            bert = BertForMaskedLM.from_pretrained('bert-base-multilingual-cased')
-        else:
-            tokenizer = None
-            bert = None
-            logger.error(f"Language {language} not supported for perplexity analysis")
-
-
-    # turn-level
-    if len(turn_df) > 0:
-
-        ## semantic similarity of current turn to previous turn of the other speaker
-        utterances_texts = utterances_filtered[measures['utterance_text']].values.tolist()
-        utterances_embeddings = sentence_encoder.encode(utterances_texts)
-        similarity_matrix = cosine_similarity(utterances_embeddings)
-
-        sentence_tangeniality1_list = []
-        sentence_tangeniality2_list = []
-        perplexity_list = []
-        perplexity_5_list = []
-        perplexity_11_list = []
-        perplexity_15_list = []
-        turn_to_turn_tangeniality_list = []
-        for i in range(len(utterances_filtered)):
-            row = utterances_filtered.iloc[i]
-            current_speaker = row[measures['speaker_label']]
-
-            if current_speaker == speaker_label:
-                phrases_texts = row[measures['phrases_texts']]
-                utterance_text = row[measures['utterance_text']]
-
-                if len(row[measures['words_texts']]) < min_coherence_turn_length:
-                    sentence_tangeniality1_list.append(np.nan)
-                    sentence_tangeniality2_list.append(np.nan)
-                    perplexity_list.append(np.nan)
-                    perplexity_5_list.append(np.nan)
-                    perplexity_11_list.append(np.nan)
-                    perplexity_15_list.append(np.nan)
-                    turn_to_turn_tangeniality_list.append(np.nan)
-                    continue
-                
-                sentence_tangeniality1, sentence_tangeniality2, perplexity, perplexity_5, perplexity_11, perplexity_15 = calculate_phrase_tangeniality(
-                    phrases_texts, utterance_text, sentence_encoder, bert, tokenizer
-                )
-                
-                sentence_tangeniality1_list.append(sentence_tangeniality1)
-                sentence_tangeniality2_list.append(sentence_tangeniality2)
-                perplexity_list.append(perplexity)
-                perplexity_5_list.append(perplexity_5)
-                perplexity_11_list.append(perplexity_11)
-                perplexity_15_list.append(perplexity_15)
-
-                if i == 0:
-                    turn_to_turn_tangeniality_list.append(np.nan)
-                elif len(utterances_filtered.iloc[i-1][measures['words_texts']]) < min_coherence_turn_length:
-                    turn_to_turn_tangeniality_list.append(np.nan)
-                else:
-                    turn_to_turn_tangeniality_list.append(similarity_matrix[i, i-1])
-
-        turn_df[measures['sentence_tangeniality1']] = sentence_tangeniality1_list
-        turn_df[measures['sentence_tangeniality2']] = sentence_tangeniality2_list
-        turn_df[measures['perplexity']] = perplexity_list
-        turn_df[measures['perplexity_5']] = perplexity_5_list
-        turn_df[measures['perplexity_11']] = perplexity_11_list
-        turn_df[measures['perplexity_15']] = perplexity_15_list
-        turn_df[measures['turn_to_turn_tangeniality']] = turn_to_turn_tangeniality_list
-
-
-    # summary-level
-    if len(turn_df) > 0:
-        for measure in ['sentence_tangeniality1', 'sentence_tangeniality2', 'perplexity', 'perplexity_5', 'perplexity_11', 'perplexity_15', 'turn_to_turn_tangeniality']:
-            summ_df[measures[measure + '_mean']] = turn_df[measures[measure]].mean(skipna=True)
-            summ_df[measures[measure + '_var']] = turn_df[measures[measure]].var(skipna=True)
-
-        summ_df[measures['turn_to_turn_tangeniality_slope']] = calculate_slope(turn_df[measures['turn_to_turn_tangeniality']])
-
-    df_list = [word_df, turn_df, summ_df]
-    return df_list
 
 def calculate_file_feature(json_data, model, speakers):
     """
